Renamed file that deals with the image Mongo Collection to be more meaningful
