+++ server/main.js
@@ -1,5 +1,5 @@
 import { Meteor } from 'meteor/meteor';
-import { Images } from "../imports/api/common.js"
+import { Images } from "../imports/api/images-collection.js"
 
 Meteor.startup(() => {
   // code to run on server at startup
